fixed #10 , separarated features,handlers
