+++ src/features/dashboard/types/dashboardTypes.ts
@@ -1,0 +1,14 @@
+import { OrderList } from "@/features/orders/types/orderTypes";
+
+export interface DashboardData {
+  thisMonthEnquiries: number;
+  lastMonthEnquiries: number;
+  thisMonthActiveUsers: number;
+  lastMonthActiveUsers: number;
+  thisMonthRevenue: number;
+  lastMonthRevenue: number;
+  thisYearRevenue: number;
+  lastYearRevenue: number;
+  monthWiseRevenue: number[];
+  last15Orders: OrderList[];
+}
