+++ src/features/product/components/forms/AddProductModal.tsx
@@ -2,33 +2,13 @@ import { zodResolver } from "@hookform/resolvers/zod";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import React, { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import * as z from "zod";
-import { Product } from "../../types/userTypes";
-import { addProduct } from "../../axios/userApi";
+import {
+  Product,
+  TAddProductSchema,
+} from "@/features/product/types/productTypes";
+import { AddProductSchema } from "@/features/product/zod/productSchemas";
+import { addProduct } from "../../axios/productApi";
 
-const AddProductSchema = z.object({
-  productName: z
-    .string()
-    .min(5, {
-      message: "Product name must be grater than or equal to 5 characters.",
-    })
-    .max(100),
-  productPrice: z.number().min(0),
-  productDescription: z
-    .string()
-    .min(5, {
-      message:
-        "Product description must be grater than or equal to 5 characters.",
-    })
-    .max(100),
-  validityInMonths: z
-    .number()
-    .min(1, {
-      message: "Product validity must be the multiple of 1 month.",
-    })
-    .max(100),
-});
-type TAddProductSchema = z.infer<typeof AddProductSchema>;
 type TModalProps = {
   buttonText: string;
   title: string;
