+++ src/features/user/components/forms/AddEnquiry.tsx
@@ -2,42 +2,13 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { useForm } from "react-hook-form";
-import { TextFieldWithLabel } from "../basic/TextFieldWithLabel";
-import * as z from "zod";
+import { TextFieldWithLabel } from "@/AppComponents/basic/TextFieldWithLabel";
+
 import { zodResolver } from "@hookform/resolvers/zod";
 import { addEnquiry } from "@/features/user/axios/userApi";
-import { User } from "@/features/user/types/userTypes";
+import { TAddEnquirySchema, User } from "@/features/user/types/userTypes";
+import { AddEnquirySchema } from "../../zod/userSchemas";
 // import { DevTool } from "@hookform/devtools"
-
-const AddEnquirySchema = z.object({
-  instituteName: z
-    .string()
-    .min(5, { message: "Institute Name must contain at least 5 character(s)" })
-    .max(100),
-  ownersName: z
-    .string()
-    .min(4, { message: "Name must be greater than or equal to 4" })
-    .max(50)
-    .optional(),
-  managersName: z
-    .string()
-    .min(4, { message: "Name must be greater than or equal to 4" })
-    .max(50)
-    .optional(),
-  address: z
-    .string()
-    .min(5, { message: "Address must contain at least 5 character(s)" })
-    .max(100),
-  phone1: z.string().min(10).max(10),
-  phone2: z.string().min(10).max(10).optional(),
-
-  email: z.string().email(),
-  website: z.string().optional(),
-  description: z.string().optional(),
-  leadType: z.string().optional(),
-  leadSource: z.string().optional(),
-});
-type TAddEnquirySchema = z.infer<typeof AddEnquirySchema>;
 
 const AddEnquiry = () => {
   const router = useRouter();
