+++ src/features/user/components/lists/TrialClientsList.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 import React from "react";
 import { User } from "@/features/user/types/userTypes";
 import { BsPersonFill } from "react-icons/bs";
-import Loading from "@/features/user/components/basic/Loading";
+import Loading from "@/AppComponents/basic/Loading";
 
 const TrialClientsList = () => {
   const queryClient = useQueryClient();
@@ -69,11 +69,11 @@ const TrialClientsList = () => {
         </Link>
       </div>
       <div className="w-full p-4 overflow-y-auto bg-white border rounded-lg">
-        <div className="grid items-center justify-between grid-cols-6 p-2 my-3 font-semibold cursor-pointer">
+        <div className="lg:grid-cols-6 grid items-center justify-between grid-cols-3 p-2 my-3 font-semibold cursor-pointer">
           <span className=" col-span-2">Institute</span>
-          <span>Contact</span>
-          <span>Status</span>
-          <span>Address</span>
+          <span className="lg:block hidden">Contact</span>
+          <span className="lg:block hidden">Status</span>
+          <span className="lg:block hidden">Address</span>
           <span>Actions</span>
         </div>
         <ul>
@@ -85,7 +85,7 @@ const TrialClientsList = () => {
               // queryClient.setQueryData(["user", user.id.toString()], user);
               <li
                 key={user.id}
-                className="bg-gray-50 hover:bg-gray-100 grid items-center justify-between grid-cols-6 gap-4 p-2 my-3 rounded-lg"
+                className="bg-gray-50 hover:bg-gray-100 lg:grid-cols-6 grid items-center justify-between grid-cols-3 p-2 my-3 cursor-pointer"
               >
                 <div className="flex items-center col-span-2">
                   <div className="p-3 bg-orange-200 rounded-lg">
@@ -98,15 +98,15 @@ const TrialClientsList = () => {
                     </span>
                   </p>
                 </div>
-                <p className="sm:text-left text-right text-gray-600">
+                <p className="lg:text-left lg:block hidden text-right text-gray-600">
                   {user.email}
                   <span className="block text-sm text-gray-500">
                     Phone:{user.phone1}
                   </span>
                 </p>
-                <p className="md:flex hidden">{user.userStatus}</p>
-                <p>{user.address}</p>
-                <div className="sm:flex gap-x-10 items-center justify-between hidden">
+                <p className="lg:block hidden">{user.userStatus}</p>
+                <p className="lg:block hidden">{user.address}</p>
+                <div className="gap-x-10 flex items-center justify-between">
                   <div className="gap-x-2 flex">
                     <Link
                       href={`/user/${user.id}`}
