+++ src/features/user/types/userTypes.ts
@@ -1,37 +1,8 @@
-export interface Product {
-  id: number;
-  productName: string;
-  productPrice: number;
-  productDescription?: string;
-  validityInMonths: number;
-}
+import * as z from "zod";
+import { AddEnquirySchema, EditUserSchema } from "../zod/userSchemas";
 
-export interface ProductInOrder extends Product {
-  isSelected?: boolean;
-  discount?: number;
-  validityFrom?: string;
-  validityUntil?: string;
-}
-export interface Order {
-  id: number;
-  userId: number;
-  products: ProductInOrder[];
-  totalAmount: number;
-  totalDiscount: number;
-  payableAmount: number;
-  paidAmount: number;
-  dueAmount: number;
-  dueDate?: string;
-  orderDate: string;
-  paymentMode: string;
-  paidBy?: string;
-  receivingAccount?: string;
-}
-export interface OrderList extends Order {
-  instituteName: string;
-  phone1: string;
-}
-
+export type TAddEnquirySchema = z.infer<typeof AddEnquirySchema>;
+export type TEditUserSchema = z.infer<typeof EditUserSchema>;
 export interface User {
   id: number;
   instituteName: string;
@@ -48,17 +19,4 @@ export interface User {
   // THIS IS NOT COMING FROM USER FORM
   addedOn: string;
   userStatus: string;
-}
-
-export interface DashboardData {
-  thisMonthEnquiries: number;
-  lastMonthEnquiries: number;
-  thisMonthActiveUsers: number;
-  lastMonthActiveUsers: number;
-  thisMonthRevenue: number;
-  lastMonthRevenue: number;
-  thisYearRevenue: number;
-  lastYearRevenue: number;
-  monthWiseRevenue: number[];
-  last15Orders: OrderList[];
 }
