+++ src/mocksRest/mswServer.ts
@@ -1,5 +1,13 @@
 import { setupServer } from "msw/node";
 
-import { handlers } from "./handlers";
+import { productHandlers } from "./handlers/productHandlers";
+import { orderHandlers } from "./handlers/orderHandlers";
+import { userHandlers } from "./handlers/userHandlers";
+import { dashboardHandlers } from "./handlers/dashboardHandlers";
 
-export const mswServer = setupServer(...handlers);
+export const mswServer = setupServer(
+  ...productHandlers,
+  ...orderHandlers,
+  ...userHandlers,
+  ...dashboardHandlers,
+);
